Move rholang tests to rholang folder (from util), update CI workflow
