+++ .github/workflows/continuous-integration.yml
@@ -124,6 +124,7 @@ jobs:
           - "'casper/test:testOnly coop.rchain.casper.engine.*'"
           - "'casper/test:testOnly coop.rchain.casper.genesis.*'"
           - "'casper/test:testOnly coop.rchain.casper.merging.*'"
+          - "'casper/test:testOnly coop.rchain.casper.rholang.*'"
           - "'casper/test:testOnly coop.rchain.casper.util.*'"
           - "'casper/test:testOnly coop.rchain.casper.sync.*'"
           - REMAINDER # Do not modify/remove!
