+++ casper/src/test/scala/coop/rchain/casper/engine/Setup.scala
@@ -10,8 +10,8 @@ import coop.rchain.casper.engine.BlockRetriever.RequestState
 import coop.rchain.casper.genesis.Genesis
 import coop.rchain.casper.genesis.contracts.Validator
 import coop.rchain.casper.protocol.{CommUtil, _}
+import coop.rchain.casper.rholang.Resources.mkTestRNodeStoreManager
 import coop.rchain.casper.rholang.RuntimeManager
-import coop.rchain.casper.util.rholang.Resources.mkTestRNodeStoreManager
 import coop.rchain.casper.util.{GenesisBuilder, TestTime}
 import coop.rchain.catscontrib.TaskContrib._
 import coop.rchain.comm._
@@ -22,7 +22,7 @@ import coop.rchain.models.{BindPattern, ListParWithRandom, Par, TaggedContinuati
 import coop.rchain.p2p.EffectsTestInstances._
 import coop.rchain.rspace.RSpace
 import coop.rchain.rspace.state.instances.RSpaceStateManagerImpl
-import coop.rchain.rspace.syntax.rspaceSyntaxKeyValueStoreManager
+import coop.rchain.rspace.syntax._
 import coop.rchain.shared.Cell
 import coop.rchain.store.InMemoryStoreManager
 import fs2.concurrent.Queue
