+++ casper/src/test/scala/coop/rchain/casper/helper/BlockDagStorageFixture.scala
@@ -6,10 +6,9 @@ import coop.rchain.blockstorage.blockStore
 import coop.rchain.blockstorage.blockStore.BlockStore
 import coop.rchain.blockstorage.dag.{BlockDagStorage, IndexedBlockDagStorage}
 import coop.rchain.casper.dag.BlockDagKeyValueStorage
-import coop.rchain.casper.rholang.RuntimeManager
+import coop.rchain.casper.rholang.{Resources, RuntimeManager}
 import coop.rchain.casper.storage.RNodeKeyValueStoreManager
 import coop.rchain.casper.util.GenesisBuilder.GenesisContext
-import coop.rchain.casper.util.rholang.{Resources}
 import coop.rchain.catscontrib.TaskContrib.TaskOps
 import coop.rchain.metrics.Metrics
 import coop.rchain.metrics.Metrics.MetricsNOP
