+++ casper/src/test/scala/coop/rchain/casper/rholang/DeployerIdTest.scala
@@ -1,11 +1,10 @@
-package coop.rchain.casper.util.rholang
+package coop.rchain.casper.rholang
 
 import cats.effect.Resource
 import com.google.protobuf.ByteString
 import coop.rchain.casper.helper.TestNode
-import coop.rchain.casper.rholang.RuntimeManager
+import coop.rchain.casper.rholang.Resources._
 import coop.rchain.casper.util.GenesisBuilder.{buildGenesis, buildGenesisParameters}
-import coop.rchain.casper.util.rholang.Resources._
 import coop.rchain.casper.util.{ConstructDeploy, ProtoUtil}
 import coop.rchain.crypto.PrivateKey
 import coop.rchain.crypto.signatures.Secp256k1
@@ -13,8 +12,8 @@ import coop.rchain.models.Expr.ExprInstance.GBool
 import coop.rchain.models.rholang.implicits._
 import coop.rchain.models.{GDeployerId, Par}
 import coop.rchain.p2p.EffectsTestInstances.LogicalTime
-import coop.rchain.shared.{Base16, Log}
 import coop.rchain.shared.scalatestcontrib.effectTest
+import coop.rchain.shared.{Base16, Log}
 import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
 import org.scalatest.{FlatSpec, Matchers}
