+++ casper/src/test/scala/coop/rchain/casper/rholang/InterpreterUtilTest.scala
@@ -1,21 +1,20 @@
-package coop.rchain.casper.util.rholang
+package coop.rchain.casper.rholang
 
-import cats.effect.{Concurrent, Sync}
+import cats.effect.Concurrent
 import cats.syntax.all._
 import com.google.protobuf.ByteString
 import coop.rchain.blockstorage.blockStore.BlockStore
 import coop.rchain.blockstorage.dag.{BlockDagStorage, DagRepresentation}
-import coop.rchain.casper.{CasperShardConf, CasperSnapshot, OnChainCasperState}
 import coop.rchain.casper.helper._
 import coop.rchain.casper.protocol._
-import coop.rchain.casper.rholang.{InterpreterUtil, RuntimeManager}
-import coop.rchain.casper.util.GenesisBuilder.{buildGenesis, buildGenesisParameters}
-import coop.rchain.casper.util.RSpaceUtil._
 import coop.rchain.casper.rholang.InterpreterUtil._
-import coop.rchain.casper.util.rholang.Resources.mkRuntimeManager
+import coop.rchain.casper.rholang.Resources.mkRuntimeManager
 import coop.rchain.casper.rholang.RuntimeManager.StateHash
 import coop.rchain.casper.rholang.types.SystemDeploy
+import coop.rchain.casper.util.GenesisBuilder.{buildGenesis, buildGenesisParameters}
+import coop.rchain.casper.util.RSpaceUtil._
 import coop.rchain.casper.util.{ConstructDeploy, GenesisBuilder}
+import coop.rchain.casper.{CasperShardConf, CasperSnapshot, OnChainCasperState}
 import coop.rchain.crypto.signatures.Signed
 import coop.rchain.metrics
 import coop.rchain.metrics.{Metrics, NoopSpan, Span}
@@ -24,15 +23,13 @@ import coop.rchain.models.PCost
 import coop.rchain.models.Validator.Validator
 import coop.rchain.p2p.EffectsTestInstances.LogStub
 import coop.rchain.rholang.interpreter.SystemProcesses.BlockData
-import coop.rchain.shared.Time
-import coop.rchain.shared.{Log, LogSource}
 import coop.rchain.shared.scalatestcontrib._
+import coop.rchain.shared.{Log, LogSource, Time}
 import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
 import org.scalatest._
 
 import scala.concurrent.duration.FiniteDuration
-import scala.util.Random
 
 class InterpreterUtilTest
     extends FlatSpec
